+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # 0.12.0
 
+- streaming response type
 - add CLI to generate api-def from openapi spec
 - add `reconfigure` method to `Api` & `Endpoint` to allow for changing options after creation
 - support for path params surrounded by curly braces, e.g. `/users/{id}`
 - add `updateBody` to `RequestContext` to allow for changing the body in middleware
+- `Endpoint.computePath` is replaced by `Endpoint.resolvePath`
 
 # 0.11.1